Move visitable interface to valuetype.ts
