+++ libs/language-server/src/lib/ast/wrappers/value-type/index.ts
@@ -3,7 +3,6 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 
 export * from './primitive';
-export * from './visitors';
 
 export * from './atomic-valuetype';
 export * from './valuetype';
