--- libs/language-server/src/lib/ast/wrappers/value-type/visitors/index.ts
@@ -1,7 +1,0 @@
-// SPDX-FileCopyrightText: 2023 Friedrich-Alexander-Universitat Erlangen-Nurnberg
-//
-// SPDX-License-Identifier: AGPL-3.0-only
-
-// eslint-disable-next-line import/no-cycle
-export * from './valuetype-visitor';
-export * from './visitable-valuetype';
